Rename worker event factory services
