+++ config/services.yaml
@@ -84,13 +84,13 @@ services:
     Psr\Http\Client\ClientInterface:
         class: GuzzleHttp\Client
 
-    App\Services\WorkerEventFactory\:
-        resource: '../src/Services/WorkerEventFactory/'
-        tags: ['foo.factory']
+    App\Services\WorkerEventFactory\EventHandler\:
+        resource: '../src/Services/WorkerEventFactory/EventHandler/'
+        tags: ['app.worker_event_factory.event_handler']
 
-    App\Services\WorkerEventFactory:
+    App\Services\WorkerEventFactory\WorkerEventFactory:
         arguments:
-            $handlers: !tagged_iterator foo.factory
+            $handlers: !tagged_iterator app.worker_event_factory.event_handler
 
     webignition\YamlDocumentGenerator\YamlGenerator:
         class: webignition\YamlDocumentGenerator\YamlGenerator
