+++ src/MessageDispatcher/DeliverEventMessageDispatcher.php
@@ -21,7 +21,7 @@ use App\Event\TestFailedEvent;
 use App\Event\TestPassedEvent;
 use App\Event\TestStartedEvent;
 use App\Message\DeliverEventMessage;
-use App\Services\WorkerEventFactory;
+use App\Services\WorkerEventFactory\WorkerEventFactory;
 use App\Services\WorkerEventStateMutator;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Symfony\Component\Messenger\Envelope;
