+++ src/Services/WorkerEventFactory/EventHandler/NoPayloadEventFactory.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services\WorkerEventFactory;
+namespace App\Services\WorkerEventFactory\EventHandler;
 
 use App\Entity\Job;
 use App\Entity\WorkerEvent;
