+++ tests/Functional/Services/WorkerEventFactory/EventHandler/CompilationFailedEventFactoryTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Functional\Services\WorkerEventFactory;
+namespace App\Tests\Functional\Services\WorkerEventFactory\EventHandler;
 
-use App\Services\WorkerEventFactory\CompilationFailedEventFactory;
-use App\Services\WorkerEventFactory\EventFactoryInterface;
+use App\Services\WorkerEventFactory\EventHandler\CompilationFailedEventFactory;
+use App\Services\WorkerEventFactory\EventHandler\EventFactoryInterface;
 use App\Tests\DataProvider\WorkerEventFactory\CreateFromCompilationFailedEventDataProviderTrait;
 
 class CompilationFailedEventFactoryTest extends AbstractEventFactoryTest
