+++ tests/Functional/Services/WorkerEventFactory/EventHandler/CompilationPassedEventFactoryTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Functional\Services\WorkerEventFactory;
+namespace App\Tests\Functional\Services\WorkerEventFactory\EventHandler;
 
-use App\Services\WorkerEventFactory\CompilationPassedEventFactory;
-use App\Services\WorkerEventFactory\EventFactoryInterface;
+use App\Services\WorkerEventFactory\EventHandler\CompilationPassedEventFactory;
+use App\Services\WorkerEventFactory\EventHandler\EventFactoryInterface;
 use App\Tests\DataProvider\WorkerEventFactory\CreateFromCompilationPassedEventDataProviderTrait;
 
 class CompilationPassedEventFactoryTest extends AbstractEventFactoryTest
