+++ tests/Functional/Services/WorkerEventFactory/EventHandler/JobTimeoutEventFactoryTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Functional\Services\WorkerEventFactory;
+namespace App\Tests\Functional\Services\WorkerEventFactory\EventHandler;
 
-use App\Services\WorkerEventFactory\EventFactoryInterface;
-use App\Services\WorkerEventFactory\JobTimeoutEventFactory;
+use App\Services\WorkerEventFactory\EventHandler\EventFactoryInterface;
+use App\Services\WorkerEventFactory\EventHandler\JobTimeoutEventFactory;
 use App\Tests\DataProvider\WorkerEventFactory\CreateFromJobTimeoutEventDataProviderTrait;
 
 class JobTimeoutEventFactoryTest extends AbstractEventFactoryTest
